requirements.txt
requests
websocket-client
socketio-client
python-socketio
pydantic
python-dotenv
asyncio
aiohttp
numpy
scipy
soundfile
librosa
pandas
matplotlib
seaborn
tqdm
colorama
loguru
pyyaml
jsonschema
openai
anthropic
google-generativeai
transformers
torch
torchaudio